:wrench: refactor - 팀 멤버 내 타입가드 서비스로직 분리 및 사용하지 않는 임포트 제거
